Renaming and TODO to use zod
